+++ packages/orq-rc/src/models/operations/getbudget.ts
@@ -351,7 +351,7 @@ export const GetBudgetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-10-23T06:53:41.205Z",
+    "2025-10-23T08:02:58.531Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -386,7 +386,7 @@ export const GetBudgetResponseBody$outboundSchema: z.ZodType<
   isActive: z.boolean(),
   consumption: z.lazy(() => GetBudgetConsumption$outboundSchema).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-10-23T06:53:41.205Z"))
+  updated: z.date().default(() => new Date("2025-10-23T08:02:58.531Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {